adds suport for credential cache lookup via env variable
