+++ KERBEROS.md
@@ -1,0 +1,49 @@
+# Using Kerberos with go-sql-driver with MySQL Enterprise
+
+This includes code changes to support kerberos authentication with the Enterprise edition of MySQL.
+
+There are three modes implemented for krb5 authentication, each are detailed below.
+
+## Environment Variables
+
+This is achieved through the use of environment variables that are common to the krb5 go library, and one additional variable unique to this driver.
+
+* KRB5_CONFIG (defaults to `/etc/krb5.conf`)
+* KRB5_CLIENT_KTNAME (user keytab)
+* KRB5CCNAME
+* KRB5_CC_LOOKUP_FILE
+
+## Keytab
+
+This is the typical setup where the user is running under a kerberos authenticated environment, and has a keytab and corresponding credential cache that is updated as needed.
+
+## Global Credential Cache
+
+This setup allows a specific credential cache to be used as a fallback where there isn't a keytab, and where there isn't a match in the lookup file.
+
+The credential cache (generated by `kinit`), can be specified with the environment variable:
+
+`KRB5CCNAME`
+
+## Per-Connection Credential Cache
+
+This provides a lookup file that maps a connection to a specific credential cache. This allows connection to multiple databases with different users using the appropriate authentication.
+
+`KRB5_CC_LOOKUP_FILE`
+
+```JSON
+[
+ {
+   "user": "usera",
+   "database": "dbone",
+   "address": "mysql1.mydomain.com:3306",
+   "credentialCache": "/tmp/krb5cc_1000"
+ },
+ {
+   "user": "userb",
+   "database": "dbtwo",
+   "address": "mysql2.mydomain.com:3306",
+   "credentialCache": "/tmp/krb5cc_1001"
+ }
+]
+```
